Moved battery_sensor.h and batter.cpp from board locations to respective locations in common
