+++ avionics/common/include/sensors/battery_sensor.h
